+++ src/services/redis/AccessTokenService.php
@@ -26,6 +26,7 @@ use Yii;
  *  database structure
  *    * oauth2:accessTokens:<aid> : hash (AccessToken)
  *    * oauth2:users:<uid>:accessTokens : set (AccessToken for user)
+ *    * oauth2:clients:<cid>:accessTokens : set (AccessToken for client)
  *
  * @author Philippe Gaultier <pgaultier@sweelix.net>
  * @copyright 2010-2017 Philippe Gaultier
@@ -37,7 +38,6 @@ use Yii;
  */
 class AccessTokenService extends BaseService implements AccessTokenServiceInterface
 {
-
     /**
      * @var string user namespace (collection for accesstokens)
      */
@@ -125,8 +125,12 @@ class AccessTokenService extends BaseService implements AccessTokenServiceInterf
         $values = $accessToken->getDirtyAttributes($attributes);
         $redisParameters = [$accessTokenKey];
         $this->setAttributesDefinitions($accessToken->attributesDefinition());
+        $expire = null;
         foreach ($values as $key => $value)
         {
+            if (($key === 'expiry') && ($value > 0)) {
+                $expire = $value;
+            }
             if ($value !== null) {
                 $redisParameters[] = $key;
                 $redisParameters[] = $this->convertToDatabase($key, $value);
@@ -137,6 +141,9 @@ class AccessTokenService extends BaseService implements AccessTokenServiceInterf
         if ($transaction === true) {
             try {
                 $this->db->executeCommand('HMSET', $redisParameters);
+                if ($expire !== null) {
+                    $this->db->executeCommand('EXPIREAT', [$accessTokenKey, $expire]);
+                }
                 if ($userAccessTokensKey !== null) {
                     $this->db->executeCommand('SADD', [$userAccessTokensKey, $accessTokenId]);
                 }
@@ -210,11 +217,18 @@ class AccessTokenService extends BaseService implements AccessTokenServiceInterf
             $redisUpdateParameters = [$accessTokenKey];
             $redisDeleteParameters = [$accessTokenKey];
             $this->setAttributesDefinitions($accessToken->attributesDefinition());
+            $expire = null;
             foreach ($values as $key => $value)
             {
                 if ($value === null) {
+                    if ($key === 'expiry') {
+                        $expire = false;
+                    }
                     $redisDeleteParameters[] = $key;
                 } else {
+                    if (($key === 'expiry') && ($value > 0)) {
+                        $expire = $value;
+                    }
                     $redisUpdateParameters[] = $key;
                     $redisUpdateParameters[] = $this->convertToDatabase($key, $value);
                 }
@@ -224,6 +238,11 @@ class AccessTokenService extends BaseService implements AccessTokenServiceInterf
             }
             if (count($redisUpdateParameters) > 1) {
                 $this->db->executeCommand('HMSET', $redisUpdateParameters);
+            }
+            if ($expire === false) {
+                $this->db->executeCommand('PERSIST', [$accessTokenKey]);
+            } elseif ($expire > 0) {
+                $this->db->executeCommand('EXPIREAT', [$accessTokenKey, $expire]);
             }
 
             $this->db->executeCommand('EXEC');
@@ -299,6 +318,21 @@ class AccessTokenService extends BaseService implements AccessTokenServiceInterf
     /**
      * @inheritdoc
      */
+    public function deleteAllByUserId($userId)
+    {
+        $userAccessTokensKey = $this->getUserAccessTokensKey($userId);
+        $userAccessTokens = $this->db->executeCommand('SMEMBERS', [$userAccessTokensKey]);
+        $userAccessTokenKeys = [$userAccessTokensKey];
+        foreach ($userAccessTokens as $userAccessToken) {
+            $userAccessTokenKeys[] = $this->getAccessTokenKey($userAccessToken);
+        }
+        $this->db->executeCommand('DEL', $userAccessTokenKeys);
+        return true;
+    }
+
+    /**
+     * @inheritdoc
+     */
     public function findAllByClientId($clientId)
     {
         $clientAccessTokensKey = $this->getClientAccessTokensKey($clientId);
@@ -310,6 +344,21 @@ class AccessTokenService extends BaseService implements AccessTokenServiceInterf
             }
         }
         return $accessTokens;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function deleteAllByClientId($clientId)
+    {
+        $clientAccessTokensKey = $this->getClientAccessTokensKey($clientId);
+        $clientAccessTokens = $this->db->executeCommand('SMEMBERS', [$clientAccessTokensKey]);
+        $clientAccessTokenKeys = [$clientAccessTokensKey];
+        foreach ($clientAccessTokens as $clientAccessToken) {
+            $clientAccessTokenKeys[] = $this->getAccessTokenKey($clientAccessToken);
+        }
+        $this->db->executeCommand('DEL', $clientAccessTokenKeys);
+        return true;
     }
 
     /**