+++ tests/unit/OauthRefreshTokenStorageTest.php
@@ -3,6 +3,8 @@
 namespace tests\unit;
 use OAuth2\Storage\RefreshTokenInterface;
 use sweelix\oauth2\server\exceptions\DuplicateKeyException;
+use sweelix\oauth2\server\interfaces\ClientModelInterface;
+use sweelix\oauth2\server\models\Client;
 use sweelix\oauth2\server\models\RefreshToken;
 use Yii;
 /**
@@ -30,7 +32,7 @@ class OauthRefreshTokenStorageTest extends TestCase
         $refreshToken->id = 'refreshToken1';
         $refreshToken->clientId = 'client1';
         $refreshToken->userId = 'user1';
-        $refreshToken->expiry = 1234;
+        $refreshToken->expiry = time() + 3600;
         $this->assertTrue($refreshToken->save());
 
         $insertedRefreshToken = RefreshToken::findOne('refreshToken1');
@@ -62,7 +64,7 @@ class OauthRefreshTokenStorageTest extends TestCase
         $refreshToken->id = 'refreshToken2';
         $refreshToken->clientId = 'client1';
         $refreshToken->userId = 'user1';
-        $refreshToken->expiry = 1234;
+        $refreshToken->expiry = time() + 3600;
         $refreshToken->scopes = ['basic'];
         $this->assertTrue($refreshToken->save());
 
@@ -72,7 +74,7 @@ class OauthRefreshTokenStorageTest extends TestCase
         $refreshToken->id = 'refreshToken3';
         $refreshToken->clientId = 'client1';
         $refreshToken->userId = 'user1';
-        $refreshToken->expiry = 1234;
+        $refreshToken->expiry = time() + 3600;
         $refreshToken->scopes = ['fail'];
         $this->assertFalse($refreshToken->save());
 
@@ -103,6 +105,132 @@ class OauthRefreshTokenStorageTest extends TestCase
 
     }
 
+    public function testDeleteAllByClientId()
+    {
+        $client1 = Yii::createObject('sweelix\oauth2\server\interfaces\ClientModelInterface');
+        /* @var ClientModelInterface $client1 */
+        $this->assertInstanceOf(ClientModelInterface::class, $client1);
+        $client1->id = 'client1';
+        $client1->secret = 'secret1';
+        $client1->isPublic = true;
+        $client1->grantTypes = [];
+        $client1->userId = 'uid';
+        $client1->scopes = [];
+        $client1->name = 'Test client';
+        $this->assertTrue($client1->save());
+
+        $this->populateScopes();
+
+        $refreshToken = Yii::createObject('sweelix\oauth2\server\interfaces\RefreshTokenModelInterface');
+        /* @var RefreshToken $refreshToken */
+        $this->assertInstanceOf(RefreshToken::className(), $refreshToken);
+        $refreshToken->id = 'refreshToken1';
+        $refreshToken->clientId = 'client1';
+        $refreshToken->userId = 'user1';
+        $refreshToken->expiry = time() + 3600;
+        $refreshToken->scopes = ['basic'];
+        $this->assertTrue($refreshToken->save());
+
+        $refreshToken = Yii::createObject('sweelix\oauth2\server\interfaces\RefreshTokenModelInterface');
+        /* @var RefreshToken $refreshToken */
+        $this->assertInstanceOf(RefreshToken::className(), $refreshToken);
+        $refreshToken->id = 'refreshToken2';
+        $refreshToken->clientId = 'client1';
+        $refreshToken->userId = 'user1';
+        $refreshToken->expiry = time() + 3600;
+        $refreshToken->scopes = ['basic'];
+        $this->assertTrue($refreshToken->save());
+
+        $refreshTokens = RefreshToken::findAllByClientId('client1');
+        $this->assertTrue(is_array($refreshTokens));
+        $this->assertEquals(2, count($refreshTokens));
+
+        $this->assertTrue(RefreshToken::deleteAllByClientId('client2'));
+
+        $refreshTokens = RefreshToken::findAllByClientId('client1');
+        $this->assertTrue(is_array($refreshTokens));
+        $this->assertEquals(2, count($refreshTokens));
+
+        $this->assertTrue(RefreshToken::deleteAllByClientId('client1'));
+
+        $refreshTokens = RefreshToken::findAllByClientId('client1');
+        $this->assertTrue(is_array($refreshTokens));
+        $this->assertEquals(0, count($refreshTokens));
+
+        $client = Client::findOne('client1');
+        $this->assertInstanceOf(Client::className(), $client);
+
+        $this->assertTrue(RefreshToken::deleteAllByClientId('client1'));
+
+        $refreshTokens = RefreshToken::findAllByClientId('client1');
+        $this->assertTrue(is_array($refreshTokens));
+        $this->assertEquals(0, count($refreshTokens));
+
+    }
+
+    public function testDeleteAllByUserId()
+    {
+        $client1 = Yii::createObject('sweelix\oauth2\server\interfaces\ClientModelInterface');
+        /* @var ClientModelInterface $client1 */
+        $this->assertInstanceOf(ClientModelInterface::class, $client1);
+        $client1->id = 'client1';
+        $client1->secret = 'secret1';
+        $client1->isPublic = true;
+        $client1->grantTypes = [];
+        $client1->userId = 'uid';
+        $client1->scopes = [];
+        $client1->name = 'Test client';
+        $this->assertTrue($client1->save());
+
+        $this->populateScopes();
+
+        $refreshToken = Yii::createObject('sweelix\oauth2\server\interfaces\RefreshTokenModelInterface');
+        /* @var RefreshToken $refreshToken */
+        $this->assertInstanceOf(RefreshToken::className(), $refreshToken);
+        $refreshToken->id = 'refreshToken1';
+        $refreshToken->clientId = 'client1';
+        $refreshToken->userId = 'user1';
+        $refreshToken->expiry = time() + 3600;
+        $refreshToken->scopes = ['basic'];
+        $this->assertTrue($refreshToken->save());
+
+        $refreshToken = Yii::createObject('sweelix\oauth2\server\interfaces\RefreshTokenModelInterface');
+        /* @var RefreshToken $refreshToken */
+        $this->assertInstanceOf(RefreshToken::className(), $refreshToken);
+        $refreshToken->id = 'refreshToken2';
+        $refreshToken->clientId = 'client1';
+        $refreshToken->userId = 'user1';
+        $refreshToken->expiry = time() + 3600;
+        $refreshToken->scopes = ['basic'];
+        $this->assertTrue($refreshToken->save());
+
+        $refreshTokens = RefreshToken::findAllByUserId('user1');
+        $this->assertTrue(is_array($refreshTokens));
+        $this->assertEquals(2, count($refreshTokens));
+
+        $this->assertTrue(RefreshToken::deleteAllByUserId('user2'));
+
+        $refreshTokens = RefreshToken::findAllByUserId('user1');
+        $this->assertTrue(is_array($refreshTokens));
+        $this->assertEquals(2, count($refreshTokens));
+
+        $this->assertTrue(RefreshToken::deleteAllByUserId('user1'));
+
+        $refreshTokens = RefreshToken::findAllByUserId('user1');
+        $this->assertTrue(is_array($refreshTokens));
+        $this->assertEquals(0, count($refreshTokens));
+
+        $client = Client::findOne('client1');
+        $this->assertInstanceOf(Client::className(), $client);
+
+        $this->assertTrue(RefreshToken::deleteAllByUserId('user1'));
+
+        $refreshTokens = RefreshToken::findAllByUserId('user1');
+        $this->assertTrue(is_array($refreshTokens));
+        $this->assertEquals(0, count($refreshTokens));
+
+    }
+
     public function testUpdate()
     {
         $refreshToken = Yii::createObject('sweelix\oauth2\server\interfaces\RefreshTokenModelInterface');
@@ -111,7 +239,7 @@ class OauthRefreshTokenStorageTest extends TestCase
         $refreshToken->id = 'refreshToken1';
         $refreshToken->clientId = 'client1';
         $refreshToken->userId = 'user1';
-        $refreshToken->expiry = 1234;
+        $refreshToken->expiry = time() + 3600;
         $this->assertTrue($refreshToken->save());
 
         $insertedRefreshToken = RefreshToken::findOne('refreshToken1');
@@ -127,7 +255,7 @@ class OauthRefreshTokenStorageTest extends TestCase
         $refreshToken->id = 'refreshToken1';
         $refreshToken->clientId = 'client1';
         $refreshToken->userId = 'user1';
-        $refreshToken->expiry = 1234;
+        $refreshToken->expiry = time() + 3600;
         $this->assertTrue($refreshToken->save());
 
         $insertedRefreshToken->id = 'refreshToken1';
@@ -144,7 +272,7 @@ class OauthRefreshTokenStorageTest extends TestCase
         $refreshToken->id = 'refreshToken1';
         $refreshToken->clientId = 'client1';
         $refreshToken->userId = 'user1';
-        $refreshToken->expiry = 1234;
+        $refreshToken->expiry = time() + 3600;
         $this->assertTrue($refreshToken->save());
 
         $insertedRefreshToken = RefreshToken::findOne('refreshToken1');
@@ -158,13 +286,14 @@ class OauthRefreshTokenStorageTest extends TestCase
         $storage = Yii::createObject('sweelix\oauth2\server\storage\OauthStorage');
         /* @var RefreshTokenInterface $storage */
         $this->assertInstanceOf(RefreshTokenInterface::class, $storage);
-        $this->assertTrue($storage->setRefreshToken('refreshToken1', 'client1', 'user1', 1250));
+        $expiry = time()+3600;
+        $this->assertTrue($storage->setRefreshToken('refreshToken1', 'client1', 'user1', $expiry));
         $refreshToken = RefreshToken::findOne('refreshToken1');
         $this->assertInstanceOf(RefreshToken::className(), $refreshToken);
         $this->assertEquals('refreshToken1', $refreshToken->id);
         $this->assertEquals('client1', $refreshToken->clientId);
         $this->assertEquals('user1', $refreshToken->userId);
-        $this->assertEquals(1250, $refreshToken->expiry);
+        $this->assertEquals($expiry, $refreshToken->expiry);
 
         $refreshTokenData = $storage->getRefreshToken('refreshToken1');
         $this->assertEquals($refreshToken->id, $refreshTokenData['refresh_token']);
@@ -173,13 +302,13 @@ class OauthRefreshTokenStorageTest extends TestCase
         $this->assertEquals($refreshToken->expiry, $refreshTokenData['expires']);
 
         $this->populateScopes();
-        $this->assertTrue($storage->setRefreshToken('refreshToken2', 'client1', 'user1', 1250, 'basic'));
+        $this->assertTrue($storage->setRefreshToken('refreshToken2', 'client1', 'user1', $expiry, 'basic'));
 
         $refreshTokenData = $storage->getRefreshToken('refreshToken2');
         $this->assertEquals('refreshToken2', $refreshTokenData['refresh_token']);
         $this->assertEquals('client1', $refreshTokenData['client_id']);
         $this->assertEquals('user1', $refreshTokenData['user_id']);
-        $this->assertEquals(1250, $refreshTokenData['expires']);
+        $this->assertEquals($expiry, $refreshTokenData['expires']);
         $this->assertEquals('basic', $refreshTokenData['scope']);
 
         $this->assertTrue($storage->unsetRefreshToken('refreshToken2'));